Extensive Refactoring and cleaning up of main code files.
I strived to comply with PEP8 wherever possible. There are still a lot of
missing docstrings.

In the configuration creation utility, I refactored the user input parts
out of the file creation. I will probably also take the definition of the
config file format out of the code and into a variable.